\name{laser}
\docType{data}
\alias{laser}
\title{laser data}
\description{
The quality characteristic of a laser device is its operating current. When the 
operating current reaches at a predefined threshold level 10, this device is 
considered to be failed. There are 15 tested units. The measured frequency of this current is 250 hours, and
the experiment was terminated at 4000 hours.
}
%%\usage{laser}
%%\format{A vector containing 141 observations.}
%%\source{World Almanac and Book of Facts, 1975, page 406.}
\references{
Meeker, W. Q. and Escobar, L. A. (1998), \emph{Statistical Methods for Reliability Data}.   New York: John Wiley and Sons.
}
%%\keyword{datasets}