\name{run.gui}
\alias{run.gui}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The graphical user interface of integrated degradation model
}
\description{
run.gui() starts the graphical user interface of integrated degradation model.

}
\usage{
run.gui()
}


\references{

Cheng, Y. S. and Peng, C. Y. (2012), "Integrated Degradation Models in R Using iDEMO." \emph{Journal of Statistical Software}, \bold{49(2)}, 1-22.

Peng, C. Y. and Tseng, S. T. (2009), "Mis-specification Analysis of Linear Degradation Models." \emph{IEEE Transactions on Reliability}, \bold{58}, 444-455.

}

